Close NetcdfFileWriter that's in define mode (fixes #718)
Previously, attempting to close a NetcdfFileWriter that was still in define mode would cause an exception. That possibility clearly hadn't been planned for and I had to make several different changes to make the operation work!

Tests will follow in a separate commit, on the 5.0.0 branch.

+++ cdm/src/main/java/ucar/nc2/NetcdfFileWriter.java
@@ -924,6 +924,10 @@ public class NetcdfFileWriter implements Closeable {
     spiw.close();
 
     File prevFile = new File(location);
+    if (!prevFile.exists()) {
+      return;
+    }
+
     File tmpFile = new File(location + ".tmp");
     if (tmpFile.exists()) {
       boolean ok = tmpFile.delete();
